+++ CHANGELOG.md
@@ -7,11 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- feat(syslogexporter): add mTLS support [#980]
+
 ### Changed
 
 - feat(sumologicextension): deprecate `install_token` [#969]
+- feat(syslogexporter): remove adding additional structure data in syslog exporter [#975]
+- feat(syslogexporter): change TLS configuration options to opentelemetry configtls [#983]
 
 [#969]: https://github.com/SumoLogic/sumologic-otel-collector/pull/969
+[#975]: https://github.com/SumoLogic/sumologic-otel-collector/pull/975
+[#980]: https://github.com/SumoLogic/sumologic-otel-collector/pull/980
+[#983]: https://github.com/SumoLogic/sumologic-otel-collector/pull/983
 [unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.71.0-sumo-0...main
 
 ## [v0.71.0-sumo-0]